+++ iree/hal/local/task_command_buffer.c
@@ -14,7 +14,7 @@
 
 #include "iree/hal/local/task_command_buffer.h"
 
-#include "iree/base/debugging.h"
+#include "iree/base/internal/debugging.h"
 #include "iree/base/tracing.h"
 #include "iree/hal/local/local_descriptor_set_layout.h"
 #include "iree/hal/local/local_executable.h"
