+++ iree/task/executor.c
@@ -14,7 +14,7 @@
 
 #include "iree/task/executor.h"
 
-#include "iree/base/debugging.h"
+#include "iree/base/internal/debugging.h"
 #include "iree/base/math.h"
 #include "iree/task/executor_impl.h"
 #include "iree/task/task_impl.h"
