+++ iree/task/pool.c
@@ -14,7 +14,7 @@
 
 #include "iree/task/pool.h"
 
-#include "iree/base/debugging.h"
+#include "iree/base/internal/debugging.h"
 #include "iree/base/math.h"
 
 // Minimum byte size of a block in bytes, including the tasks as well as the
