+++ iree/task/task.c
@@ -16,7 +16,7 @@
 
 #include <stdio.h>
 
-#include "iree/base/debugging.h"
+#include "iree/base/internal/debugging.h"
 #include "iree/task/task_impl.h"
 
 //==============================================================================
