+++ iree/task/topology.c
@@ -18,7 +18,7 @@
 #include <cpuinfo.h>
 #include <stdio.h>
 
-#include "iree/base/debugging.h"
+#include "iree/base/internal/debugging.h"
 #include "iree/base/math.h"
 #include "iree/base/tracing.h"
 #include "iree/task/tuning.h"
